Add tests for 100% coverage LaunchAngleInterpolator
Use Dependency Injection for ConfigLoader to make the LaunchAngleInterpolator more testable. Doing so, we can modify it to accept a delegate or an interface for loading configurations. This allows us to inject mock behaviors during testing.

This involves modifying the constructor to accept a delegate for loading configuration. Now we can simulate different scenarios without relying on actual file I/O.

For something this simple, its a little overkill but I think its valuable to have this template code in our codebase for more complex tests in the future